Moved DLCUtil to core
